+++ roadgeometrypoint.cpp
@@ -1,6 +1,6 @@
 #include "roadgeometrypoint.h"
 
-RoadGeometryPoint::RoadGeometryPoint(QPointF position, float medianWidth, float lateralMedianOffset)
+RoadGeometryPoint::RoadGeometryPoint(QPointF position, qreal medianWidth, qreal lateralMedianOffset)
     : position_{ position }
     , medianWidth_{ medianWidth }
     , medianOffset_{ lateralMedianOffset }
