+++ roadgeometrypoint.h
@@ -4,17 +4,17 @@
 class RoadGeometryPoint
 {
     QPointF position_;
-    float medianWidth_;
-    float medianOffset_;
+    qreal medianWidth_;
+    qreal medianOffset_;
 public:
-    RoadGeometryPoint(QPointF position, float medianWidth = 0.0f, float medianOffset = 0.0f);
+    RoadGeometryPoint(QPointF position, qreal medianWidth = 0.0f, qreal medianOffset = 0.0f);
 
     QPointF position() const { return position_; }
-    float medianWidth() const { return medianWidth_; }
-    float medianOffset() const { return medianOffset_; }
+    qreal medianWidth() const { return medianWidth_; }
+    qreal medianOffset() const { return medianOffset_; }
 
     void position(QPointF& position) { position_ = position; }
-    void medianWidth(float width) { medianWidth_ = width; }
-    void medianOffset(float offset) { medianOffset_ = offset; }
+    void medianWidth(qreal width) { medianWidth_ = width; }
+    void medianOffset(qreal offset) { medianOffset_ = offset; }
 
 };
